Add TODO: CustomTimerDuration in Settings commented code start
Started working on CustomTimerDuration through settings. This belongs to one of the scenarios, but maybe more important to get the other scenarios working for now to start self-exploration/self-testing. This is why i added these code-pieces as commented code, this needs work still (especially logic). I looked at the codepieces for the intervall settings.