FocusAssist check workaround: Balloon Tooltip, see description
There is no officially supported API from Microsoft to check the status of Focus assist, which is important to be turned on for a focus session as to prevent the user being disturbed by notifications (and corresponding sounds). I could have gone with unofficial C++ code from Github where the author wrote in the Readme that the API used can change/break at any time in the future, but this is a more elegant, always supported workaround: If Focus Assist is turned on ('Priority only' or 'Alarms only'), the user will not see the balloon. If it is not turned on, the user will see the balloon to turn on focus assist. The icon itself serves as a session indicator. The only downside is, with focus assist active, the notification center will still display the message, but it will be disposed if the session ends/ is stopped.