Add checking for total missed slack messages at the end of session
At session stop, it will count all the missed slack messages in the associated slack workspace from the bot-token, and only the channels that the bot had been added to to watch over, and inform the user on how many messages were received during the session in this workspace. The bot at the same time for demonstration purposes currently postet a message every 10 seconds in the #general channel, so there are missed messages to report on. Currently when testing, the client.GetChannelListAsync() takes a super long time and has not returned (i cancelled execution after a while) need to check this later again