Add Settings Customized Email Reply
In the Settings, the user can customize the automatic email reply and activate it (of course will only send customized reply if automatic reply is activated). If the user deactivated the customized reply, it will simply send the standard reply.
PS since i set the replymessage at the start of a session, currently after changing the customized reply message, the a new session needs to be started / restarted for it to take effect.