Add checks for email and slack and adapt queries
Email checks if Office 365 Tracker is enabled, then it will summarize received emails. If response enabled in settings, it will additionally automatically respond.
Slack is configured over json file. If default values from generated file, slack is seen as inactive.
If botAuthToken is added, it is seen as active, can connect to the corresponding workspace, and will summarize received messages of the public channels in that workspace where the focussession app has been added to.
If memberId is overwritten, it will check on usermention in channel (by memberId) and automatically respond,