Add Pause/Continue and Shutdown behavior
When Personal Analytics is paused during an active session, it is handled as a manual stop by the user, since focus sessions are meant to be done at one continual piece. If the user tries to start a session while Personal Analytics is paused, the user is informed that Personal Analytics needs to be unpaused first. If Personal Analytics is shutdown while an active session, it will be logged in the db. It will also be logged if the session was terminated thorugh pausing or shutdown.