Change default reply in open session
Sender does not need to know for how long the user is focusing on a task. This makes sense in case of a closed Session, to show the sender for how long the recipient will not receive the message. But the time duration display in case of an open session does not give the sender any valuable information since we do not know for how long the user intends to run the session