More subtle windowFlagged behavior
If the user continues to navigate within the same application (like opening different emails in Thunderbird or switching channels within Slack), the application gets flagged only once. If the user furthermore clicks Yes, meaning the application is task-related, it is counted as allowed Application for this session and will not be flagged again during that session. If the user clicks no, and switches to another application, and later on back again, it should be flagged again to remind the user to focus on the task.